Modularise code dedicated to database
